Use NoCure to avoid having to do scripting
Add NoCure to seedsprout worm and electric snail so that they don't show
up as possible fungal infection cures. This is a little less flexible
than the Harmony-based approach, which created an allowlist for usable
corpses (this approach is effectively creating a deny list). However, it
requires zero scripting and is less likely to break in future versions
of Qud.